# optimization for reading hive file  when all columns to read are partition keys