--- a/tests/integration/test_hive_query/test.py
+++ b/tests/integration/test_hive_query/test.py
@@ -375,7 +375,7 @@ def test_cache_read_bytes(started_cluster):
     for i in range(10):
         result = node.query(
             """
-    SELECT day, count(*) FROM default.demo_parquet_1 group by day order by day settings input_format_parquet_allow_missing_columns = true
+    SELECT * FROM default.demo_parquet_1 settings input_format_parquet_allow_missing_columns = true
             """
         )
         node.query("system flush logs")