### Changelog category (leave one):
- Performance improvement


### Changelog entry (a user-readable short description of the changes that goes to CHANGELOG.md):

When all the columns to read are partition keys, construct columns by the file's row number without real reading the hive file.

With following test sqls
```SQL
CREATE TABLE test_day_hour
(
    `rip` Nullable(String),
    `rtime` Nullable(String),
    `stattime` Nullable(Int64),
    `msgid` Nullable(String),
    `msgtype` Nullable(Int64),
    `day` String,
    `hour` String
)
ENGINE = Hive('thrift://xxxx', 'db', 'test_table')
PARTITION BY (day, hour);

SELECT
    day,
    count(*)
FROM test_day_hour
WHERE (day = '2022-05-01') AND (hour >= '20')
GROUP BY day

```

we have performance results.
* before optimize
```
1 row in set. Elapsed: 2.202 sec. Processed 99.18 million rows, 2.39 GB (45.04 million rows/s., 1.09 GB/s.)
```
* after optimize
```
1 row in set. Elapsed: 0.899 sec. Processed 99.18 million rows, 73.14 KB (110.33 million rows/s., 81.36 KB/s.)
```


> Information about CI checks: https://clickhouse.com/docs/en/development/continuous-integration/