@lgbo-ustc 2022-05-12
> > Integration tests (release, actions) [1/2] — fail: 1
> 
> ```
> test_hive_query/test.py::test_cache_read_bytes
> ```

It is the case that all required colums are partition keys, and no data was read from file.  

Fixed it.