@lgbo-ustc 2022-05-13
The failures don't seem to be related to the hive part ?