@kssenii 2022-05-19
> ClickHouse special build check (actions) — 6/7 artifact groups are OK

```
std=gnu++2a -MD -MT src/CMakeFiles/dbms.dir/Backups/BackupUtils.cpp.o -MF src/CMakeFiles/dbms.dir/Backups/BackupUtils.cpp.o.d -o src/CMakeFiles/dbms.dir/Backups/BackupUtils.cpp.o -c ../src/Backups/BackupUtils.cpp
May 17 02:25:15 /build/build_docker/../src/Backups/BackupUtils.cpp:369:22: error: 'auto &' can be declared as 'const auto &' [readability-qualified-auto,-warnings-as-errors]
May 17 02:25:15                 for (auto & [relative_path, backup_entry] : info.data)
May 17 02:25:15                      ^
May 17 02:25:15                      const 
May 17 02:25:15 /build/build_docker/../src/Backups/BackupUtils.cpp:377:18: error: 'auto &' can be declared as 'const auto &' [readability-qualified-auto,-warnings-as-errors]
May 17 02:25:15             for (auto & [relative_path, backup_entry] : info.data)
May 17 02:25:15                  ^
May 17 02:25:15                  const 
May 17 02:25:15 214897 warnings generated.
```

> Integration tests (release, actions) [2/2] — fail: 1

test_multiple_disks/test.py::test_concurrent_alter_move_and_drop[replicated]